Hopefully should work now

+++ src/cnn/VisDroneCNN.py
@@ -1,12 +1,11 @@
 import torch
 import torch.nn as nn
-import logging
 from src.cnn.FitParentClass import FitParentClass
 
 logger = logging.getLogger(__name__)
 
 class VisDroneCNN(FitParentClass):
-    def __init__(self, B_boxes: int = 1, C: int = 10):
+    def __init__(self, B_boxes: int = 1, C: int = 10, dtype=torch.float32):
         """
         B_boxes: number of bounding boxes per grid cell (most models only use 1)
         C: number of classes (here 10)
